Partial creation of initial functions for sending buttons to display
Created an object to hold the display values so that it can be easily
manipulated and read in each function. Created function to return which
button was pressed and started function for logic of updating the object
holding the display values. Created skeleton of function to update the
actual display.