Refactor/cleanup code and fix bug after getting result
Cleaned up the code so it is no longer a bowl of spaghetti and separated
each section into regions for improved readability.

Fixed a bug found while refactoring with the numbersA array not
resetting properly after getting a result. Previously it would only
allow one number to be entered after getting result, now should properly
allow max display size amount of numbers to be entered. Also added in a
constant for the max display size so that it is easier to change if the
display size is ever changed.